backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# AWS
boto3==1.34.25
aioboto3==12.3.0

# WebSocket
websockets==12.0
python-socketio==5.11.0

# AI/ML
anthropic==0.18.0
openai==1.10.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
transformers==4.37.2
accelerate==0.26.1
diffusers==0.26.3
numpy==1.26.3
opencv-python==4.9.0.80
Pillow==10.2.0
librosa==0.10.1
soundfile==0.12.1
scipy==1.12.0

# STT
openai-whisper==20231117
faster-whisper==0.10.0

# TTS
TTS==0.22.0
bark==0.1.4

# Avatar Animation
face-alignment==1.4.1
imageio==2.33.1
imageio-ffmpeg==0.4.9
scikit-image==0.22.0
kornia==0.7.1

# Video Processing
ffmpeg-python==0.2.0

# Celery
celery==5.3.6
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk==1.40.0

# Utilities
httpx==0.26.0
aiofiles==23.2.1
tenacity==8.2.3
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
